// frontend/src/lib/api/generated/models/QuestionAnswerItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { AnswerItem } from '../models/AnswerItem';
import type { QuestionTypeCode } from '../models/QuestionTypeCode';
import { HttpFile } from '../http/http';

export class QuestionAnswerItem {
    'id': number;
    'surveyId': number;
    'typeCode': QuestionTypeCode;
    'question': string;
    'orderNumber': number;
    'answers'?: Array<AnswerItem>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": ""
        },
        {
            "name": "surveyId",
            "baseName": "survey_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "typeCode",
            "baseName": "type_code",
            "type": "QuestionTypeCode",
            "format": ""
        },
        {
            "name": "question",
            "baseName": "question",
            "type": "string",
            "format": ""
        },
        {
            "name": "orderNumber",
            "baseName": "order_number",
            "type": "number",
            "format": ""
        },
        {
            "name": "answers",
            "baseName": "answers",
            "type": "Array<AnswerItem>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return QuestionAnswerItem.attributeTypeMap;
    }

    public constructor() {
    }
}