// frontend/src/lib/api/generated/models/SurveyResponseResultItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class SurveyResponseResultItem {
    'questionId': number;
    'answersIds'?: Array<number>;
    'answerText'?: string | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "questionId",
            "baseName": "question_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "answersIds",
            "baseName": "answers_ids",
            "type": "Array<number>",
            "format": ""
        },
        {
            "name": "answerText",
            "baseName": "answer_text",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SurveyResponseResultItem.attributeTypeMap;
    }

    public constructor() {
    }
}